Remove hard-coded CC,CCC,CXX as at least two of these is defined by make/gnu-make (what uses CCC?).  I did a simple test and fedora core and win10+mingw do have sensible defaults.
This should fix #20.

As a follow-up note: It is pretty wild we never deleted 90% of this make system
we inherited from netbeans launcher.  It is much more sophisticated than what
we need here.  I am somewhat loathe to touch this as I have a lot of other
things to do.  We have also been talking over the years about replacing this
with something else (mruby/rust/svm).