Do not add CLASSPATH to the module path
Treating the CLASSPATH as modules will create automatic modules for all
entries that are not actually modules. This can lead to several problems,
e.g. the automatic module name is determined by the name of the jar or
directory, which fails when one component is not a valid java identifier.
Another issue is that several automatic modules might export the same
package, which is not allowed and leads to an error.

To fix these issues, we use the module path just for jruby libraries and
pass the CLASSPATH using -classpath, as both can be combined.

This is identical to the fix for jruby.bash in jruby/jruby (20e4278) [1].

[1]: https://github.com/jruby/jruby/pull/5728