Handle add-options appropriately across JRuby versions
For JRuby versions earlier than 9.2.1, we do not specify add-opens
flags because we do not export a module to use.

For JRuby versions from 9.2.1 through 9.2.9, we hardcode some
add-opens for org.jruby.dist, because these versions did not ship
a .module_opts file.

For JRuby versions 9.2.10 and higher we use .module_opts file.