Fix 2 issues in windows launcher when JAVA_HOME missing
1. We searched for 'java' and not 'java.exe' so it never found
anything.  Also we hard-bailed in that code vs continue on to later
searches.
2. Registry querying was only working on 32 bit windows OSes.  Mingw32
is the only thing we use to compile the launcher so we get a 32 bit
binary.  Most OSes are 64 bit now and to query the registry we are
required to pass in an option specifying that.Z