Default to module support on and disable on 9.2-
Defaulting to module support off meant master would build a binary
that doesn't open up appropriate packages, which is supported
by the last 11 JRuby releases. This change defaults to on so that
the default build will add those flags. It disables module flags
if the JRuby version installing the gem is 9.2.0 or lower.