Stop modifying RubyGems defaults
This file was added to switch JRuby's RubyGems back to using
normal shebang lines (rather than env shebangs needed for the
shell script) but had the effect of making that JRuby install need
a native launcher permanently. Rather than switching off env
shebangs, we leave them as is so the native launcher and shell
launcher can be interchangeable.